CI: Add markdownlint rule to check for undefined reference links
By default, markdownlint detects missing link targets
in reference links written in the "collapsed syntax" `[foo][]`,
but not if they use the "shortcut syntax" [foo].

Since markdownlint 0.31.0, the rule MD052/reference-links-images
now supports a `shortcut_syntax` boolean parameter
that allows opting-in to checking shortcut syntax links as well.

This requires markdownlint-cli v0.37.0 or above,
but since our GitHub Actions setup fetches the latest available version,
we already have access to it in the CI pipeline,
so we only need to add the rule and parameter
in the markdownlint configuration file.